code/requirements.txt
fastapi
uvicorn
PyPDF2
openai
python-multipart
flask
pandas
openpyxl
xgboost
joblib
scikit-learn
torch
datasets
transformers
accelerate>=0.26.0
pymupdf
python-docx
pytesseract